third_party/WebKit/LayoutTests/inspector/sass/test-ast-scss-1.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script src="./sass-test.js"></script>
<style>
pre {
    font-family: monospace;
}
</style>
<script>

function test()
{
    InspectorTest.evaluateInPage("getSASS()", onSASS);

    function onSASS(result)
    {
        InspectorTest.parseSCSS("", result.value)
            .then(InspectorTest.dumpAST)
            .then(InspectorTest.validateASTRanges)
            .catch(console.error.bind(console))
            .then(InspectorTest.completeTest);
    }
}

function getSASS()
{
    return document.querySelector(".snippet").textContent;
}

</script>
</head>

<body onload="runTest()">
<p>
Verifies SCSS AST in case of simple nesting and CSS properties.
</p>
<pre class="snippet">
div {
    /* display: flex; */
    color: $my-color;
    background-color: blue;
    /* position: absolute; */

    .className {
        padding: 10px 0 0 10px;
        /* font-family: "Times New Roman"; */
        background-image: url(assets/no-image-set.png);

        #test::before {
            margin: 10px 10px;
            content: "test me";
            /* border: 1px solid black; */
        }
    }
}
</pre>
</body>
</html>
