third_party/WebKit/LayoutTests/inspector/sass/test-ast-scss-1-expected.txt

Verifies SCSS AST in case of simple nesting and CSS properties.

div {
    /* display: flex; */
    color: $my-color;
    background-color: blue;
    /* position: absolute; */

    .className {
        padding: 10px 0 0 10px;
        /* font-family: "Times New Roman"; */
        background-image: url(assets/no-image-set.png);

        #test::before {
            margin: 10px 10px;
            content: "test me";
            /* border: 1px solid black; */
        }
    }
}
=== AST === 
rule 0
    selector 0: "div"
        range: {"startLine":0,"startColumn":0,"endLine":0,"endColumn":3}
    property 0
        name: "display"
            range: {"startLine":1,"startColumn":7,"endLine":1,"endColumn":14}
        value: "flex"
            range: {"startLine":1,"startColumn":16,"endLine":1,"endColumn":20}
        range: {"startLine":1,"startColumn":4,"endLine":1,"endColumn":24}
        disabled: true
    property 1
        name: "color"
            range: {"startLine":2,"startColumn":4,"endLine":2,"endColumn":9}
        value: "$my-color"
            range: {"startLine":2,"startColumn":11,"endLine":2,"endColumn":20}
        range: {"startLine":2,"startColumn":4,"endLine":2,"endColumn":21}
        disabled: false
    property 2
        name: "background-color"
            range: {"startLine":3,"startColumn":4,"endLine":3,"endColumn":20}
        value: "blue"
            range: {"startLine":3,"startColumn":22,"endLine":3,"endColumn":26}
        range: {"startLine":3,"startColumn":4,"endLine":3,"endColumn":27}
        disabled: false
    property 3
        name: "position"
            range: {"startLine":4,"startColumn":7,"endLine":4,"endColumn":15}
        value: "absolute"
            range: {"startLine":4,"startColumn":17,"endLine":4,"endColumn":25}
        range: {"startLine":4,"startColumn":4,"endLine":4,"endColumn":29}
        disabled: true
rule 1
    selector 0: ".className"
        range: {"startLine":6,"startColumn":4,"endLine":6,"endColumn":14}
    property 0
        name: "padding"
            range: {"startLine":7,"startColumn":8,"endLine":7,"endColumn":15}
        value: "10px 0 0 10px"
            range: {"startLine":7,"startColumn":17,"endLine":7,"endColumn":30}
        range: {"startLine":7,"startColumn":8,"endLine":7,"endColumn":31}
        disabled: false
    property 1
        name: "font-family"
            range: {"startLine":8,"startColumn":11,"endLine":8,"endColumn":22}
        value: ""Times New Roman""
            range: {"startLine":8,"startColumn":24,"endLine":8,"endColumn":41}
        range: {"startLine":8,"startColumn":8,"endLine":8,"endColumn":45}
        disabled: true
    property 2
        name: "background-image"
            range: {"startLine":9,"startColumn":8,"endLine":9,"endColumn":24}
        value: "url(assets/no-image-set.png)"
            range: {"startLine":9,"startColumn":26,"endLine":9,"endColumn":54}
        range: {"startLine":9,"startColumn":8,"endLine":9,"endColumn":55}
        disabled: false
rule 2
    selector 0: "#test::before"
        range: {"startLine":11,"startColumn":8,"endLine":11,"endColumn":21}
    property 0
        name: "margin"
            range: {"startLine":12,"startColumn":12,"endLine":12,"endColumn":18}
        value: "10px 10px"
            range: {"startLine":12,"startColumn":20,"endLine":12,"endColumn":29}
        range: {"startLine":12,"startColumn":12,"endLine":12,"endColumn":30}
        disabled: false
    property 1
        name: "content"
            range: {"startLine":13,"startColumn":12,"endLine":13,"endColumn":19}
        value: ""test me""
            range: {"startLine":13,"startColumn":21,"endLine":13,"endColumn":30}
        range: {"startLine":13,"startColumn":12,"endLine":13,"endColumn":31}
        disabled: false
    property 2
        name: "border"
            range: {"startLine":14,"startColumn":15,"endLine":14,"endColumn":21}
        value: "1px solid black"
            range: {"startLine":14,"startColumn":23,"endLine":14,"endColumn":38}
        range: {"startLine":14,"startColumn":12,"endLine":14,"endColumn":42}
        disabled: true
======
Ranges OK.
